Fix graph container permissions and graph acl - #1576

+++ pandora_console/include/functions_container.php
@@ -208,7 +208,16 @@ function folder_get_sibling($sibling) {
 }
 
 function ui_toggle_container($code, $name, $title = '', $hidden_default = true, $return = false, $group , $id_container, $parent = false) {
-	// Generate unique Id
+	
+  global $config;
+  $report_r = check_acl ($config['id_user'], 0, "RR");
+  $report_w = check_acl ($config['id_user'], 0, "RW");
+  $report_m = check_acl ($config['id_user'], 0, "RM");
+  $access = ($report_r == true) ? 'RR' : (($report_w == true) ? 'RW' : (($report_m == true) ? 'RM' : 'RR'));
+  
+  html_debug($report_w);
+  
+  // Generate unique Id
 	$uniqid = uniqid('');
 	
 	// Options
@@ -277,12 +286,16 @@ function ui_toggle_container($code, $name, $title = '', $hidden_default = true, 
     $data = array();
     $data[0] = '<a href="javascript:" id="tgl_ctrl_'.$uniqid.'">' . html_print_image ($original, true, array ("title" => $title, "id" => "image_".$uniqid)) . '&nbsp;&nbsp;<b>'.$name.'</b></a>';
     $data[1] = ui_print_group_icon($group,true);
+    if($report_r && $report_w){
     $data[2] = '<a href="index.php?sec=reporting&sec2=godmode/reporting/create_container&edit_container=1&id='.
         $id_container .'">'.html_print_image("images/config.png", true).'</a>';
-    if ($id_container !== '1'){
-        $data[2] .= '&nbsp;&nbsp;&nbsp;&nbsp' .'<a href="index.php?sec=reporting&sec2=godmode/reporting/graph_container&delete_container=1&id='
-            .$id_container.'" onClick="if (!confirm(\''.__('Are you sure?').'\'))
-            return false;">' . html_print_image("images/cross.png", true, array('alt' => __('Delete'), 'title' => __('Delete'))) . '</a>';    
+      }
+    if($report_r && $report_w && $report_m){
+      if ($id_container !== '1'){
+          $data[2] .= '&nbsp;&nbsp;&nbsp;&nbsp' .'<a href="index.php?sec=reporting&sec2=godmode/reporting/graph_container&delete_container=1&id='
+              .$id_container.'" onClick="if (!confirm(\''.__('Are you sure?').'\'))
+              return false;">' . html_print_image("images/cross.png", true, array('alt' => __('Delete'), 'title' => __('Delete'))) . '</a>';    
+      }
     }
     $table->data[] = $data;
     $table->rowclass[] = '';
